+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Hide ACF admin menu
+
 ### 1.2.4: October 28th, 2019
 * Upgraded Mpdf to version 8.x
 * Upgraded Advanced Custom Fields to version 5.x